refactor(pyroscope): extract DebugInfoClient to leaf package
Moves the struct + Upload method out of package write into a new
write/debuginfoclient package, breaking the cycle that prevented
reporter tests from reusing the production type. Reporter tests now
use debuginfoclient.Client instead of a duplicated test stub.
